* ext/dl/cfunc.c (rb_dlcfunc_inspect): suppress warnings by   rb_sprintf.
* ext/dl/cptr.c (rb_dlptr_s_malloc): suppress warnings.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@28148 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/dl/cptr.c
@@ -186,7 +186,7 @@ static VALUE
 rb_dlptr_s_malloc(int argc, VALUE argv[], VALUE klass)
 {
     VALUE size, sym, obj;
-    int   s;
+    long s;
     freefunc_t f;
 
     switch (rb_scan_args(argc, argv, "11", &size, &sym)) {
